Move hooks and slices related to UI to UI folder
