+++ src/battle/pages/BattlesPage/BattlesPage.tsx
@@ -4,7 +4,7 @@ import useBattles from "../../hooks/useBattles";
 import Loader from "../../../ui/components/Loader/Loader";
 import Paginator from "../../../ui/components/Paginator/Paginator";
 import BattlesList from "../../components/BattlesList/BattlesList";
-import useLoading from "../../../hooks/useLoading";
+import useLoading from "../../../ui/hooks/useLoading";
 import "./BattlesPage.css";
 
 const BattlesPage: React.FC = () => {
