+++ src/main.tsx
@@ -7,7 +7,7 @@ import AppRouter from "./router/AppRouter";
 import "@fontsource-variable/roboto/index.css";
 import "@fontsource-variable/roboto-slab/index.css";
 import "@fontsource-variable/museomoderno/index.css";
-import "./ui/styles/styles.css";
+import "./styles/styles.css";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
