+++ src/ui/hooks/__tests__/endLoading.test.tsx
@@ -1,12 +1,12 @@
 import { renderHook } from "@testing-library/react";
-import store from "../../store/store";
+import store from "../../../store/store";
 import { Provider } from "react-redux";
 import useLoading from "../useLoading";
 import { act } from "react";
 
-describe("Given the startLoading method from useLoading hook", () => {
+describe("Given the endLoading method from useLoading hook", () => {
   describe("When it is called", () => {
-    test("Then it should start the loading", async () => {
+    test("Then it should end the loading", async () => {
       const wrapper = ({ children }: { children: React.ReactNode }) => (
         <Provider store={store}>{children}</Provider>
       );
@@ -14,12 +14,12 @@ describe("Given the startLoading method from useLoading hook", () => {
       const { result } = renderHook(() => useLoading(), { wrapper });
 
       await act(() => {
-        result.current.startLoading();
+        result.current.endLoading();
       });
 
       const isLoading = result.current.loadingState.isLoading;
 
-      expect(isLoading).toBe(true);
+      expect(isLoading).toBe(false);
     });
   });
 });
