+++ src/ui/hooks/useLoading.ts
@@ -3,7 +3,7 @@ import {
   endLoadingActionCreator,
   startLoadingActionCreator,
 } from "../slices/loadingSlice";
-import { useAppDispatch, useAppSelector } from "../store/hooks";
+import { useAppDispatch, useAppSelector } from "../../store/hooks";
 import type { UseLoadingStructure } from "./types";
 
 const useLoading = (): UseLoadingStructure => {
